Extract session handling to SessionHandler
